Resolved the cppcheck style warnings
cppcheck is a popular tool that performs static code analysis
outputting warnings if constructs are error-prone or do not make
sense. Currently nanosvg produces many warnings, though, making
spotting real errors in a project harder than necessary.

This commit resolves the following warings:

  nanosvg.h:829: style: Variable 'data' is reassigned a value before the old one has been used.
  nanosvg.h:2907: style: Variable 'fp' is reassigned a value before the old one has been used.

The cause of these warnings is that both variables are set to
NULL and then assigned a different value to - which is no bug
but adds an unnecessary assignment to the code.