Fix drawing artifacts due to rounding errors.
This patch fixes issue #130.

I have seen that there are already some for fixing those graphic glitches, but in my opinion they are not exactly correct.
Increasing the value of NSVG__FIXSHIFT just hides the real cause of the problem.
In my opinion, there is a mathematical error into the conversion from the floating point value to fixed point.
Here, the code is using floorf() for converting the floating point value to integer, but I think that this is wrong because it does not take care of the remaining valid digits of that number.
Rounding to the nearest high or low integer value discards the value of the valid digits not included into the fractional part of the fixed point number.
As result, this error is accumulated on each step and, at the end, the limit of the integral does not converge to the expected value.
That's why increasing NSVG__FIXSHIFT apparently fixes the issue, but in my opinion a better solution is to change how the integer value is converted and keeping NSVG__FIXSHIFT as it is, at least for now.

In other words, floorf() needs to be replaced with roundf().
This patch has fixed the test case included into the issue #130.

According to the requests, a new function nsvg__roundf() has been introduced, for supporting compilation under C89 and MSVC6.